change iptable rules and watching mnp from broker

+++ pkg/controller/multinetworkpolicy/fullsync.go
@@ -4,6 +4,7 @@ import (
 	"crypto/sha256"
 	"encoding/base32"
 	"fmt"
+	mcsv1a1 "sigs.k8s.io/mcs-api/pkg/apis/v1alpha1"
 	"strconv"
 	"strings"
 	"sync"
@@ -11,7 +12,6 @@ import (
 	"github.com/pkg/errors"
 	octopusv1alpha1 "github.com/pmlproject9/octopus/pkg/apis/octopus.io/v1alpha1"
 	"github.com/pmlproject9/octopus/pkg/constants"
-	listersoctopusv1alpha1 "github.com/pmlproject9/octopus/pkg/generated/listers/octopus.io/v1alpha1"
 	"github.com/pmlproject9/octopus/pkg/ipset"
 	submv1 "github.com/submariner-io/submariner/pkg/apis/submariner.io/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -20,6 +20,7 @@ import (
 	listerscorev1 "k8s.io/client-go/listers/core/v1"
 	"k8s.io/klog/v2"
 	utilexec "k8s.io/utils/exec"
+	listermcsv1a1 "sigs.k8s.io/mcs-api/pkg/client/listers/apis/v1alpha1"
 )
 
 func (ctr *Controller) FullSync(wg *sync.WaitGroup) {
@@ -38,16 +39,22 @@ func (ctr *Controller) FullSync(wg *sync.WaitGroup) {
 		case <-ctr.ctx.Done():
 			klog.V(4).Info("shutting down the fullsync grouting")
 			return
-
 		}
 	}
 }
 
 func (ctr *Controller) fullSync() {
-	ctr.ensureDefaultIPtables()
-
-	ctr.ensureAllMultiNetworkPolicy()
-
+	klog.Info("applying default iptables rule")
+	if err := ctr.ensureDefaultIPtables(); err == nil {
+		klog.Info("success to set up default iptables")
+	} else {
+		klog.Errorf("failed to set up default iptables: err %v", err)
+	}
+	klog.Info("applying all multinetworkpolicy rules")
+	if success := ctr.ensureAllMultiNetworkPolicy(); success {
+		klog.Info("success to set up all multinetworkpolicy iptables")
+	}
+	klog.Info("cleaning up stale iptables and ipset ")
 	ctr.cleanUpStaleIptablesAndIPSet()
 }
 
@@ -65,7 +72,9 @@ func (ctr *Controller) ReFlushServicesCidr(wg *sync.WaitGroup) {
 		case <-ctr.reFlushChan:
 			err := ctr.reFlushServiceCidrIPSet()
 			if err != nil {
-				klog.Errorf("error to update ipset %s : %v", constants.IPSetServicePrefix, err)
+				klog.Errorf("error to update ipset %s : %v", constants.IPSetAllServicesCidr, err)
+			} else {
+				klog.Infof("success update ipset %s", constants.IPSetAllServicesCidr)
 			}
 		case <-ctr.ctx.Done():
 			klog.V(4).Info("shutting down the reflush ipset grouting")
@@ -100,32 +109,49 @@ func (ctr *Controller) reFlushServiceCidrIPSet() error {
 	return err
 }
 
-func (ctr *Controller) ensureDefaultIPtables() {
+func (ctr *Controller) ensureDefaultIPtables() error {
 	err := ctr.iptablesRunner.CreateChainIfNotExist(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSFORWARD)
 	if err != nil {
+		fmt.Println("err:", err)
 		klog.Errorf("error to create iptables chain %s", constants.IPTablesChainOCTOPUSFORWARD)
+		return err
 	}
 	err = ctr.iptablesRunner.CreateChainIfNotExist(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSMNPFORWARD)
 	if err != nil {
 		klog.Errorf("error to create iptables chain %s", constants.IPTablesChainOCTOPUSMNPFORWARD)
+		return err
 	}
 	forwardToOctopusForward := []string{"-m", "comment", "--comment", "OCTOPUS", "-j", "OCTOPUS-FORWARD"}
 	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, "FORWARD", 1, forwardToOctopusForward); err != nil {
 		klog.Errorf("error to insert iptables chain %s rule %s : %v", "FORWARD", strings.Join(forwardToOctopusForward, " "), err)
+		return err
+	}
+	responseAccept := []string{"-m", "conntrack", "--ctstate", "RELATED,ESTABLISHED", "-j", "ACCEPT"}
+	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSFORWARD, 1, responseAccept); err != nil {
+		klog.Errorf("error to insert iptables chain %s rule %s : %v", constants.IPTablesChainOCTOPUSMNPFORWARD, strings.Join(responseAccept, " "), err)
+		return err
+	}
+	invalidDrop := []string{"-m", "conntrack", "--ctstate", "INVALID", "-j", "DROP"}
+	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSFORWARD, 2, invalidDrop); err != nil {
+		klog.Errorf("error to insert iptables chain %s rule %s : %v", constants.IPTablesChainOCTOPUSMNPFORWARD, strings.Join(invalidDrop, " "), err)
+		return err
 	}
 	forwardToMNPForward := []string{"-m", "set", "--match-set", constants.IPSetAllServicesCidr, "dst", "-j", constants.IPTablesChainOCTOPUSMNPFORWARD}
-	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSFORWARD, 1, forwardToMNPForward); err != nil {
+	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSFORWARD, 3, forwardToMNPForward); err != nil {
 		klog.Errorf("error to insert iptables chain %s rule %s : %v", "FORWARD", strings.Join(forwardToMNPForward, " "), err)
+		return err
 	}
 	MNPStart := []string{"-m", "comment", "--comment", "start of multicluster policies", "-j", "MARK", "--set-xmark", "0x0/0x200000"}
 	if err = ctr.iptablesRunner.InsertUnique(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSMNPFORWARD, 1, MNPStart); err != nil {
 		klog.Errorf("error to insert iptables chain %s rule %s: %v", constants.IPTablesChainOCTOPUSMNPFORWARD, strings.Join(MNPStart, " "), err)
+		return err
 	}
-
 	MNPEnd := []string{"-m", "comment", "--comment", "Drop if no multi network policies passed packed", "-m", "mark", "--mark", "0x0/0x200000", "-j", "DROP"}
 	if err = ctr.iptablesRunner.InsertAtEnd(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSMNPFORWARD, MNPEnd); err != nil {
 		klog.Errorf("error to insert iptables chain %s rule %s: %v", constants.IPTablesChainOCTOPUSMNPFORWARD, strings.Join(MNPEnd, " "), err)
+		return err
 	}
+	return nil
 }
 
 func (ctr *Controller) cleanUpStaleIptablesAndIPSet() {
@@ -138,22 +164,14 @@ func (ctr *Controller) cleanUpStaleIptablesAndIPSet() {
 		if !ok {
 			klog.Errorf("error to convert multinetworkpolicy")
 		}
-		mnpNameHash := getMultiNetworkPolicyNameHash(mnp.Namespace, mnp.Name)
-		activeIPset[constants.IPSetPodPrefix+mnpNameHash] = true
-		activeIPset[constants.IPSetServicePrefix+mnpNameHash] = true
-		activeIPtablesChain[constants.IPTablesChainOCTOPUSFWPREFIX+mnpNameHash] = true
-	}
-	ipsetList, err := ctr.ipsetRunner.ListIPSets()
-	if err != nil {
-		klog.Errorf("error to list ipset: %v", err)
-	}
-	for _, ips := range ipsetList {
-		if strings.Contains(ips, constants.IPSetPodPrefix) || strings.Contains(ips, constants.IPSetServicePrefix) {
-			if _, ok := activeIPset[ips]; !ok {
-				ctr.ipsetRunner.Destroy(ips)
-			}
+		if entries, _ := ctr.ListPodIPsOfMnp(mnp); len(entries) > 0 {
+			mnpNameHash := getMultiNetworkPolicyNameHash(mnp.Namespace, mnp.Name)
+			activeIPset[constants.IPSetPodPrefix+mnpNameHash] = true
+			activeIPset[constants.IPSetServicePrefix+mnpNameHash] = true
+			activeIPtablesChain[constants.IPTablesChainOCTOPUSFWPREFIX+mnpNameHash] = true
 		}
 	}
+
 	iptablesChains, err := ctr.iptablesRunner.ListChains(constants.IPTablesFilter)
 	if err != nil {
 		klog.Errorf("error list iptables chain from filter")
@@ -184,6 +202,18 @@ func (ctr *Controller) cleanUpStaleIptablesAndIPSet() {
 			}
 		}
 	}
+
+	ipsetList, err := ctr.ipsetRunner.ListIPSets()
+	if err != nil {
+		klog.Errorf("error to list ipset: %v", err)
+	}
+	for _, ips := range ipsetList {
+		if strings.Contains(ips, constants.IPSetPodPrefix) || strings.Contains(ips, constants.IPSetServicePrefix) {
+			if _, ok := activeIPset[ips]; !ok {
+				ctr.ipsetRunner.Destroy(ips)
+			}
+		}
+	}
 }
 
 func getMultiNetworkPolicyNameHash(namespace string, name string) string {
@@ -192,27 +222,37 @@ func getMultiNetworkPolicyNameHash(namespace string, name string) string {
 	return encode[:16]
 }
 
-func (ctr *Controller) ensureAllMultiNetworkPolicy() {
-
+func (ctr *Controller) ensureAllMultiNetworkPolicy() bool {
+	setUpSuccess := true
 	for _, mnpObj := range ctr.multiNetworkPolicyLister.List() {
 		mnp, ok := mnpObj.(*octopusv1alpha1.MultiNetworkPolicy)
 		if !ok {
 			klog.Errorf("error to convert multinetworkpolicy\n")
+			setUpSuccess = false
 		}
 		if err := ctr.ensureMultiNetworkPolicy(mnp); err != nil {
-			klog.Errorf("%v", err)
+			klog.Errorf("error to set up all multinetworkpolicy %v", err)
+			setUpSuccess = false
 		}
 	}
+	return setUpSuccess
 }
 
 func (ctr *Controller) ensureMultiNetworkPolicy(mnp *octopusv1alpha1.MultiNetworkPolicy) error {
-	err := ctr.ensureMultiNetworkPolicyIPsets(mnp)
+	iptablesCreated, err := ctr.ensureMultiNetworkPolicyIPsets(mnp)
 	if err != nil {
 		return errors.Wrap(err, fmt.Sprintf("error to convert multinetworkpolicy %s into ipset: %v", mnp.Namespace+"/"+mnp.Name, err))
 	}
-	err = ctr.ensureMultiNetworkPolicyIPtables(mnp)
-	if err != nil {
-		return errors.Wrap(err, fmt.Sprintf("error to convert multinetworkpolicy %s into iptable rule:%v", mnp.Namespace+"/"+mnp.Name, err))
+	if iptablesCreated {
+		err = ctr.ensureMultiNetworkPolicyIPtables(mnp)
+		if err != nil {
+			return errors.Wrap(err, fmt.Sprintf("error to convert multinetworkpolicy %s into iptable rule:%v", mnp.Namespace+"/"+mnp.Name, err))
+		}
+	} else {
+		err = ctr.deleteMultiNetworkPolicy(mnp)
+		if err != nil {
+			return errors.Wrap(err, fmt.Sprintf("error to delete "))
+		}
 	}
 	return nil
 }
@@ -226,80 +266,53 @@ func (ctr *Controller) deleteMultiNetworkPolicy(mnp *octopusv1alpha1.MultiNetwor
 	ipsetServiceName := constants.IPSetServicePrefix + mnpNameHash
 	iptableName := constants.IPTablesChainOCTOPUSFWPREFIX + mnpNameHash
 
-	args := []string{"-m", "set", "--match-set", ipsetPodName, "src", "-j", iptableName}
-	if err := ctr.iptablesRunner.Delete(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSMNPFORWARD, args...); err != nil {
-		return errors.Wrap(err, "error to delete iptable rule")
+	if existed, err := ctr.iptablesRunner.ChainExists(constants.IPTablesFilter, iptableName); err == nil && existed {
+		args := []string{"-m", "set", "--match-set", ipsetPodName, "src", "-j", iptableName}
+		if err := ctr.iptablesRunner.Delete(constants.IPTablesFilter, constants.IPTablesChainOCTOPUSMNPFORWARD, args...); err != nil {
+			return errors.Wrap(err, "error to delete iptable rule")
+		}
+		if err := ctr.iptablesRunner.DeleteChainDirect(constants.IPTablesFilter, iptableName); err != nil {
+			return errors.Wrap(err, "error to delete chain "+iptableName)
+		}
 	}
-	if err := ctr.iptablesRunner.DeleteChainDirect(constants.IPTablesFilter, iptableName); err != nil {
-		return errors.Wrap(err, "error to delete chain "+iptableName)
-	}
-	if err := ctr.ipsetRunner.Destroy(ipsetPodName); err != nil {
+
+	if err := ctr.ipsetRunner.DestroyIfExist(ipsetPodName); err != nil {
 		return errors.Wrap(err, "error delete ipset "+ipsetPodName)
 	}
-	if err := ctr.ipsetRunner.Destroy(ipsetServiceName); err != nil {
+	if err := ctr.ipsetRunner.DestroyIfExist(ipsetServiceName); err != nil {
 		return errors.Wrap(err, "error to delete ipset "+ipsetServiceName)
 	}
 	return nil
 }
 
-func (ctr *Controller) ensureMultiNetworkPolicyIPsets(mnp *octopusv1alpha1.MultiNetworkPolicy) error {
+func (ctr *Controller) ensureMultiNetworkPolicyIPsets(mnp *octopusv1alpha1.MultiNetworkPolicy) (bool, error) {
 	ctr.ipsetMutex.Lock()
 	defer ctr.ipsetMutex.Unlock()
 
+	podEntries, _ := ctr.ListPodIPsOfMnp(mnp)
+	klog.Infof("multinetworkpolicy %s/%s pod entries %v", mnp.Namespace, mnp.Name, podEntries)
+	if len(podEntries) <= 0 {
+		return false, nil
+	}
+
 	mnpHashName := getMultiNetworkPolicyNameHash(mnp.Namespace, mnp.Name)
-	podSelector, _ := metav1.LabelSelectorAsSelector(&mnp.Spec.PodSelector)
 	podIPsetName := constants.IPSetPodPrefix + mnpHashName
 	serviceIPsetName := constants.IPSetServicePrefix + mnpHashName
-
-	pods, _ := ctr.ListPodsByNamesapceAndLabels(mnp.Namespace, podSelector)
-	podEntries := make([]string, 0)
-	for _, pod := range pods {
-		if isNetPolActionable(pod) {
-
-		}
-		podEntries = append(podEntries, pod.Status.PodIP)
-	}
 	if err := ctr.ipsetRunner.ReFlush(ipset.New(podIPsetName, ipset.HashNet, nil), podEntries); err != nil {
 		klog.Errorf("error to reflush ipset %s: %v", podIPsetName, err)
 	}
-	servicesEntriesMap := make(map[string]bool)
-	for _, selectors := range mnp.Spec.Egress.Allow {
-		if selectors.NamespaceSelector != nil {
-			nsSelector, _ := metav1.LabelSelectorAsSelector(selectors.NamespaceSelector)
-			nss, _ := ctr.ListNamespaceByLabels(nsSelector)
-			for _, ns := range nss {
-				services, _ := ctr.ListServiceSyncByLabels(ns.Name, labels.Everything())
-				for _, service := range services {
-					if service.Labels[constants.LabelSourceClusterID] == ctr.cluterID {
-						continue
-					}
-					for _, ip := range service.Spec.IPs {
-						servicesEntriesMap[ip] = true
-					}
-				}
-			}
-		}
-		if selectors.ServiceSelector != nil {
-			serviceSelector, _ := metav1.LabelSelectorAsSelector(selectors.ServiceSelector)
-			services, _ := ctr.ListServiceSyncByLabels(mnp.Namespace, serviceSelector)
-			for _, service := range services {
-				if service.Labels[constants.LabelSourceClusterID] == ctr.cluterID {
-					continue
-				}
-				for _, ip := range service.Spec.IPs {
-					servicesEntriesMap[ip] = true
-				}
-			}
-		}
+
+	serviceEntries, err := ctr.ListServiceIPsOfMnp(mnp)
+	if err != nil {
+		klog.Errorf("error to list service IPs of mnp %s err: %v", mnp.Name+mnp.Namespace, err)
 	}
-	servicesEntries := make([]string, 0)
-	for k := range servicesEntriesMap {
-		servicesEntries = append(servicesEntries, k)
-	}
-	if err := ctr.ipsetRunner.ReFlush(ipset.New(serviceIPsetName, ipset.HashNet, nil), servicesEntries); err != nil {
+	klog.Infof("multinetworkpolicy %s/%s service entries %v", mnp.Namespace, mnp.Name, serviceEntries)
+
+	if err := ctr.ipsetRunner.ReFlush(ipset.New(serviceIPsetName, ipset.HashNet, nil), serviceEntries); err != nil {
 		klog.Errorf("error to reflush ipset %s: %v", serviceIPsetName, err)
 	}
-	return nil
+
+	return true, nil
 }
 
 func (ctr *Controller) ensureMultiNetworkPolicyIPtables(mnp *octopusv1alpha1.MultiNetworkPolicy) error {
@@ -318,8 +331,8 @@ func (ctr *Controller) ensureMultiNetworkPolicyIPtables(mnp *octopusv1alpha1.Mul
 	}
 	args = []string{"-m", "set", "--match-set", constants.IPSetServicePrefix + mnpHashName, "dst"}
 	argsEnd := []string{"-j", "ACCEPT"}
-	if len(mnp.Spec.Egress.Ports) > 0 {
-		ports := mnp.Spec.Egress.Ports
+	if len(mnp.Spec.Filter.Ports) > 0 {
+		ports := mnp.Spec.Filter.Ports
 
 		for _, port := range ports {
 			argsCopy := []string{}
@@ -368,13 +381,86 @@ func (ctr *Controller) ListNamespaceByLabels(label labels.Selector) (ret []*core
 	return ns, nil
 }
 
-func (ctr *Controller) ListServiceSyncByLabels(namespace string, label labels.Selector) (ret []*octopusv1alpha1.ServiceSync, err error) {
-	serviceSyncLister := listersoctopusv1alpha1.NewServiceSyncLister(ctr.serviceSyncLister)
-	serviceSyncs, err := serviceSyncLister.ServiceSyncs(namespace).List(label)
+func (ctr *Controller) ListServiceImportByLabels(namespace string, label labels.Selector) (ret []*mcsv1a1.ServiceImport, err error) {
+	serviceImportLister := listermcsv1a1.NewServiceImportLister(ctr.serviceImportLister)
+	serviceImports, err := serviceImportLister.ServiceImports(namespace).List(label)
 	if err != nil {
 		return nil, errors.Wrap(err, "error to list servicesync")
 	}
-	return serviceSyncs, err
+	return serviceImports, err
+}
+
+func (ctr *Controller) ListPodIPsOfMnp(mnp *octopusv1alpha1.MultiNetworkPolicy) (ret []string, err error) {
+	podSelector, err := metav1.LabelSelectorAsSelector(&mnp.Spec.PodSelector)
+	if err != nil {
+		return nil, errors.Wrap(err, "podselector")
+	}
+	pods, _ := ctr.ListPodsByNamesapceAndLabels(mnp.Namespace, podSelector)
+	podEntries := make([]string, 0)
+	for _, pod := range pods {
+		if isNetPolActionable(pod) {
+			podEntries = append(podEntries, pod.Status.PodIP)
+		}
+	}
+	return podEntries, nil
+}
+
+func (ctr *Controller) ListServiceIPsOfMnp(mnp *octopusv1alpha1.MultiNetworkPolicy) ([]string, error) {
+	serviceIPsMap := make(map[string]bool)
+
+	for _, peer := range mnp.Spec.Filter.Allow {
+		namespaces := make([]string, 0)
+		if peer.NamespaceSelector != nil {
+			nsSelector, err := metav1.LabelSelectorAsSelector(peer.NamespaceSelector)
+			if err != nil {
+				return nil, errors.Wrap(err, "namesapceselector")
+			}
+			nss, _ := ctr.ListNamespaceByLabels(nsSelector)
+			for _, ns := range nss {
+				namespaces = append(namespaces, ns.Name)
+			}
+		} else {
+			namespaces = append(namespaces, mnp.Namespace)
+		}
+
+		var serviceSelector labels.Selector
+		if peer.ServiceSelector != nil {
+			var err error
+			serviceSelector, err = metav1.LabelSelectorAsSelector(peer.ServiceSelector)
+			if err != nil {
+				return nil, errors.Wrap(err, "serviceselector")
+			}
+		} else {
+			serviceSelector = labels.Everything()
+		}
+		services, _ := ctr.ListServiceImportByLabels(constants.SubmarinerOperator, serviceSelector)
+		for _, service := range services {
+			namespace := service.Labels[constants.LabelSourceNamespace]
+			if clusterID := service.Labels[constants.LabelSourceClusterID]; clusterID == ctr.cluterID {
+				continue
+			}
+			existed := false
+			for _, ns := range namespaces {
+				if ns == namespace {
+					existed = true
+					break
+				}
+			}
+			if !existed {
+				continue
+			}
+			for _, ip := range service.Spec.IPs {
+				serviceIPsMap[ip] = true
+			}
+		}
+	}
+
+	serviceIPs := make([]string, 0)
+	for ip, _ := range serviceIPsMap {
+		serviceIPs = append(serviceIPs, ip)
+	}
+	return serviceIPs, nil
+
 }
 
 func isNetPolActionable(pod *corev1.Pod) bool {
